fix(tab5): resolve 16-bit effect ID chain, preset corruption, and UI bugs
BUG-1/2: Effect cycling and name display broken after K1 migrated to 16-bit
hex effect IDs (0x1100+). Promote effectId to uint16_t through the full chain:
WsMessageRouter zone.status handler, WebSocketClient.sendZoneEffect signature,
ParameterMap EFFECT_MAX raised from 99→187, main.cpp lookupEffectName and
name cache fast path via s_k1EffectName, NameLookup.h forward declaration.

BUG-3: Preset system silently truncated effect IDs >255 and checksum covered
only bytes 0–39, leaving zone effectId high bytes unprotected. Fix: rename
ZonePresetConfig.effectId→effectIdLow, add setZoneEffectId()/getZoneEffectId()
helpers using reservedFuture[1..4], move checksum field to offset 62 (covers
full struct), bump CURRENT_VERSION to 2, add v1→v2 NVS migration in
PresetStorage.load(), update PresetManager capture/apply paths accordingly.

CR-4: enum class PresetSlotState defined identically in PresetSlotWidget.h and
PresetBankWidget.h — ODR violation. Extract to shared PresetTypes.h and include
from both widgets.

CR-7: ButtonHandler index≥8 catch-all silently discarded Unit-B button presses
that should drive toggleZoneMode() and toggleSpeedPaletteMode(). Add explicit
dispatch table for indices 8, 9, 11, 13, 15 before the catch-all.

CR-9: sendZoneBrightness() and sendZoneBlend() lacked the isConnected() guard
and queue fallback that sendZoneSpeed/sendZonePalette already had. Bring them
into parity so disconnected sends are queued rather than silently dropped.

Build: pio run -e tab5 → SUCCESS (68.9% flash, 3.3% RAM)

+++ tab5-encoder/src/network/WebSocketClient.h
@@ -129,7 +129,7 @@ public:
     // ========================================================================
 
     void sendZoneEnable(bool enable);
-    void sendZoneEffect(uint8_t zoneId, uint8_t effectId);
+    void sendZoneEffect(uint8_t zoneId, uint16_t effectId);
     void sendZoneBrightness(uint8_t zoneId, uint8_t value);
     void sendZoneSpeed(uint8_t zoneId, uint8_t value);
     void sendZonePalette(uint8_t zoneId, uint8_t paletteId);
